requirements.txt
ibm-watsonx-ai>=1.0.8
langchain>=0.2.5
langchain-ibm>=0.1.3
langchain-community>=0.2.5
chromadb>=0.5.0
gradio>=4.37.2
pypdf>=4.2.0
huggingface-hub>=0.23.4
fpdf2>=2.7.8
pytest>=8.2.0
